requirements.txt
pydantic
numpy
huggingface-hub
pillow
scikit-learn
opencv-python
pandas
mediapipe
tensorflow[and-cuda]
tf_keras_vis
pyarrow
streamlit
